requirements.txt
absl-py==1.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==25.1.0
blinker==1.7.0
boto3==1.37.18
botocore==1.37.18
brotlipy==0.7.0
cachetools==5.2.0
certifi==2021.10.8
cffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/croot-2p6q7e2r/cffi_1642701115063/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
coloredlogs==15.0.1
conda==4.12.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///opt/concourse/worker/volumes/live/0cf65687-72c2-4e51-5322-1b3a8c3aa6f5/volume/conda-package-handling_1649105815927/work
contourpy==1.3.0
cryptography @ file:///opt/concourse/worker/volumes/live/a9ca20dd-e1fe-4ed3-754e-0db64082a5d2/volume/cryptography_1639414576087/work
cycler==0.12.1
decorator==5.2.0
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-cli==0.0.5
filelock==3.16.1
Flask==3.0.2
Flask-Cors==5.0.0
flatbuffers==22.10.26
fonttools==4.54.1
frappe==0.0.1
frappe-bench==5.22.6
fsspec==2024.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
groq==0.11.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.7.0
honcho==1.1.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==5.0.0
importlib_resources==6.4.5
inquirerpy==0.3.4
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==2.11.0
kiwisolver==1.4.7
libclang==14.0.6
lxml==5.3.0
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.9.2
mdurl==0.1.2
moviepy==2.1.2
mpmath==1.3.0
mysql-connector-python==9.1.0
narwhals==1.24.2
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.19.2
opencv-python==4.11.0.86
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==21.3
pandas==2.2.3
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.0
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///opt/concourse/worker/volumes/live/112288ac-9cb0-4e73-768b-13baf4ca6419/volume/pysocks_1605305820043/work
pytesseract==0.3.13
python-crontab==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
python-pptx==1.0.2
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel-yaml-conda @ file:///opt/concourse/worker/volumes/live/e81cf0fe-611a-498e-6e69-a7320057c1ac/volume/ruamel_yaml_1616016689696/work
s3transfer==0.11.4
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
selenium==4.29.0
semantic-version==2.8.5
sentry-sdk==2.18.0
setproctitle==1.3.4
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.2
streamlit==1.41.1
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
transformers==4.48.2
trio==0.29.0
trio-websocket==0.12.2
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.32.0
uvloop==0.21.0
wandb==0.18.7
watchfiles==0.24.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.0
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.1.2
XlsxWriter==3.2.0
yt-dlp==2025.2.19
zipp==3.10.0
